backend/requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
tensorflow== 2.20.0
google-cloud-storage
python-dotenv
pydantic
pydantic-settings
python-multipart